Add generic tests for checking the get_response_json method and remove them from provider scripts
