+++ src/cc_catalog_airflow/dags/provider_api_scripts/common/test_requester.py
@@ -1,5 +1,8 @@
 import requests
 import time
+from unittest.mock import patch, MagicMock
+
+import pytest
 
 from common import requester
 
@@ -28,3 +31,81 @@ def test_get_handles_exception(monkeypatch):
 
     dq = requester.DelayedRequester(1)
     dq.get('https://google.com/')
+
+
+def test_get_response_json_retries_with_none_response():
+    dq = requester.DelayedRequester(1)
+    with patch.object(
+            dq,
+            'get',
+            return_value=None
+    ) as mock_get:
+        with pytest.raises(Exception):
+            assert dq.get_response_json(
+                'https://google.com/',
+                retries=2,
+                query_params={}
+            )
+
+    assert mock_get.call_count == 3
+
+
+def test_get_response_json_retries_with_non_ok():
+    dq = requester.DelayedRequester(1)
+    r = requests.Response()
+    r.status_code = 504
+    r.json = MagicMock(return_value={'batchcomplete': ''})
+    with patch.object(
+            dq,
+            'get',
+            return_value=r
+    ) as mock_get:
+        with pytest.raises(Exception):
+            assert dq.get_response_json(
+                'https://google.com/',
+                retries=2,
+                query_params={}
+            )
+
+    assert mock_get.call_count == 3
+
+
+def test_get_response_json_retries_with_error_json():
+    dq = requester.DelayedRequester(1)
+    r = requests.Response()
+    r.status_code = 200
+    r.json = MagicMock(return_value={'error': ''})
+    with patch.object(
+            dq,
+            'get',
+            return_value=r
+    ) as mock_get:
+        with pytest.raises(Exception):
+            assert dq.get_response_json(
+                'https://google.com/',
+                retries=2,
+                query_params={}
+            )
+
+    assert mock_get.call_count == 3
+
+
+def test_get_response_json_returns_response_json_when_all_ok():
+    dq = requester.DelayedRequester(1)
+    expect_response_json = {'batchcomplete': ''}
+    r = requests.Response()
+    r.status_code = 200
+    r.json = MagicMock(return_value=expect_response_json)
+    with patch.object(
+            dq,
+            'get',
+            return_value=r
+    ) as mock_get:
+        actual_response_json = dq.get_response_json(
+            'https://google.com/',
+            retries=2,
+            query_params={}
+        )
+
+    assert mock_get.call_count == 1
+    assert actual_response_json == expect_response_json
